increasing U - bug fixes

+++ 5_General/general.cpp
@@ -767,7 +767,7 @@ void test_flip_spins(){
     /* flip spins */
     flip_spin_v(lattice, t, l);
 }
-void sweep_lattice(LaGenMatComplex& lattice, const int lattice_size, const int time_size, const double U, const double lambda, const double delta_tau, const int iterations, int& acceptance, int& rejection){
+void sweep_lattice(LaGenMatComplex& lattice, const int lattice_size, const int time_size, const double U, const double lambda, const double delta_tau, const int iterations, double& acceptance, double& rejection){
 
     /* initialise everything */
     COMPLEX weightBefore;
